requirements.txt
Flask==2.2.2
requests==2.28.1
google-cloud-storage==2.7.0
gunicorn==20.1.0
Werkzeug== 2.2.3 